+++ api/src/main/java/org/allaymc/api/world/sound/PressurePlateActivateSound.java
@@ -1,0 +1,12 @@
+package org.allaymc.api.world.sound;
+
+import org.allaymc.api.block.type.BlockState;
+
+/**
+ * Sound played when a pressure plate is activated.
+ *
+ * @param blockState the pressure plate block state; the sound may depend on the block type
+ * @author daoge_cmd
+ */
+public record PressurePlateActivateSound(BlockState blockState) implements Sound {
+}