+++ server/src/main/java/org/allaymc/server/block/component/trapdoor/BlockTrapdoorBaseComponentImpl.java
@@ -1,6 +1,8 @@
 package org.allaymc.server.block.component.trapdoor;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.data.BlockFace;
+import org.allaymc.api.block.dto.Block;
 import org.allaymc.api.block.dto.PlayerInteractInfo;
 import org.allaymc.api.block.type.BlockState;
 import org.allaymc.api.block.type.BlockType;
@@ -10,6 +12,7 @@ import org.allaymc.api.world.Dimension;
 import org.allaymc.api.world.sound.DoorCloseSound;
 import org.allaymc.api.world.sound.DoorOpenSound;
 import org.allaymc.server.block.BlockPlaceHelper;
+import org.allaymc.server.block.RedstoneHelper;
 import org.allaymc.server.block.component.BlockBaseComponentImpl;
 import org.joml.Vector3ic;
 
@@ -56,5 +59,35 @@ public class BlockTrapdoorBaseComponentImpl extends BlockBaseComponentImpl {
                 isOpen ? new DoorOpenSound(clickedBlockState.getBlockState()) : new DoorCloseSound(clickedBlockState.getBlockState())
         );
         return true;
+    }
+
+    @Override
+    public void onNeighborUpdate(Block block, Block neighbor, BlockFace face) {
+        super.onNeighborUpdate(block, neighbor, face);
+
+        // Check redstone power and update trapdoor state
+        checkRedstonePower(block);
+    }
+
+    /**
+     * Checks if the trapdoor should be opened/closed based on redstone power.
+     *
+     * @param block the trapdoor block
+     */
+    protected void checkRedstonePower(Block block) {
+        int power = RedstoneHelper.getReceivedPowerLevel(block);
+
+        boolean shouldBeOpen = power > 0;
+        boolean isCurrentlyOpen = block.getPropertyValue(OPEN_BIT);
+
+        if (shouldBeOpen != isCurrentlyOpen) {
+            var dimension = block.getDimension();
+            var pos = block.getPosition();
+
+            dimension.updateBlockProperty(OPEN_BIT, shouldBeOpen, pos);
+
+            // Play sound
+            dimension.addSound(MathUtils.center(pos), shouldBeOpen ? new DoorOpenSound(block.getBlockState()) : new DoorCloseSound(block.getBlockState()));
+        }
     }
 }